Make guess-language-backward-paragraph always go to word boundary
This fixes the problem of marking the last word from previous
paragraph as incorrect if the current paragraph is in different
language.

This is caused by the following: When a change of language is
detected, the whole paragraph is flyspelled by calling flyspell-region
on the paragraph. guess-language-backward-paragraph finds the start of
the paragraph to check and in many cases (e.g. in text-mode), the
found point is on the empty line separating two paragraphs.
flyspell-region start spelling from that point by calling
flyspell-word. This function checks the word under the point or if the
point is in between words, the previous word. Since the start of the
paragraph is between words, the last word of the previous paragraph is
checked against dictionary of current paragraph.

This is fixed by ensuring that guess-language-backward-paragraph
always goes to the first non-whitespace character of the paragraph.
This causes flyspell-word to check the first word in the paragraph and
not that last word on the previous paragraph.