Make the word under point correctly highlighted after change of dictionary
When the dictionary is switched as a result of detecting different
language, the word under point remains highlighted as
incorrect (according to the previously used dictionary) until the word
is flyspelled again, for example by moving the point out of the word
and back. This is annoying.

This commit fixes that by ensuring that flyspelling the current
paragraph with the new dictionary happens after the return from
flyspell-incorrect-hook, which highlights the word under point
according to the old dictionary.